+++ src/main/java/com/imsweb/validation/ValidationEngine.java
@@ -89,8 +89,8 @@ public class ValidationEngine {
     /**
      * Engine version (used to check compatibility with the edits)
      */
-    private static final String _ENGINE_VERSION = "6.7";
-
+    private static final String _ENGINE_VERSION = "6.8";
+    
     /**
      * The different context types supported by the engine
      */